# Extend PartialTuple for structs