This aims to address the following issue:

Say we have:
```
function foo(b)
    a = 1
    f = ()->Val(a)
    f()
end
```

This infers beautifully, because the closure struct gets `Const`'ed
and thus inter-procedural constant prop takes care of it. However,
if we instead do

```
function foo(b)
    a = 1
    f = ()->(println(b); Val(a))
    f()
end
```

the best inference can tell us about this is that we get `Val`, because
the captured arguments are no longer constant. This leads to significant
inference problems for Zygote, because the backwards pass is always
specified as a closure. Thus, even if constant information is present
in the forward pass, it is often lost for the part of the function
that's the backwards pass, because it has to pass through the closure
struct. This fixes that by keeping track of field types individually.

Of course as always, there is a concern that this will increase compile times.
I see three options:
1. Always put this behind an inference parameter, as we do with aggressive_const_prop
2. Same as 1, but unconditionally enable it for closures
3. Always enable it

Thoughts?