@JeffBezanson 2019-01-21
We should try always enabling this. It doesn't seem likely to me to increase compile times much, so let's try it.