@vtjnash 2019-01-22
Unlike Jeff apparently, I assumed this would affect compilation time negatively (or at least memory usage, and thus runtime) a lot on certain subclasses of problems, for little overall benefit. Or just would make compilation less predictable and harder to explain (although, that ship probably sailed years ago). And hence had constrained the original implementation just to tuples, where the benefit seemed much more likely, and unexpected behaviors would hopefully be less likely. I have sometimes thought it would be nice to have this though...

Also, apparently nothing else in the lattice can intersect with this? I didn't realize that. I thought PartialTypeVar might be a problem for doing this, but it looks like we have a special magic constructor for those (no idea why, but that's what we have right now), so that should be safe (as long as nobody uses it—or any other of our other special types for that matter (or perhaps anything non-isbits)—from a generated function, such as a StructOfArrays-like package—but I digress).