@JeffBezanson 2019-01-22
> Also, apparently nothing else in the lattice can intersect with this?

Is that what this implements? That certainly sounds wrong... is there a bug in `⊑`?