@Keno 2019-01-22
What'd be a good measure of compile time impact? Time to run the tests (or one of them? subarray?)?