@JeffBezanson 2019-01-22
Time to precompile DataFrames, Plots, JuliaDB, time for `plot(rand(10))` the first time. Time to build Base and stdlibs, and yes subarray tests is a good one too.