@Keno 2019-01-22
A'right. I'll take some of those numbers.