@vtjnash 2019-01-22
> Is that what this implements?

Yes (it includes updates to `⊑`), but "nothing else" is a fairly small list