@Keno 2019-01-22
Alright numbers are in. Bit of a mixed back. This branch is about 3-4% slower on sysimg/subarray, slightly faster on the various precompiles and first plot. These numbers are not super stable though:
```
This branch:

Sysimage built. Summary:
Total ───────  75.344423 seconds
Base: ───────  25.989335 seconds 34.494%
Stdlibs: ────  49.353342 seconds 65.5036%

    JULIA test/subarray
Test  (Worker) | Time (s) | GC (s) | GC % | Alloc (MB) | RSS (MB)
running testset subarray...
subarray   (1) |   202.10 |  16.85 |  8.3 |   25529.19 |   761.86

Test Summary: | Pass  Total
  Overall     |  281    281
    SUCCESS

real    3m30.521s
user    3m29.025s
sys 0m3.372s

julia> @time using DataFrames
[ Info: Precompiling DataFrames [a93c6f00-e57d-5684-b7b6-d8193f3e46c0]
 23.501432 seconds (8.65 M allocations: 465.927 MiB, 0.82% gc time)

julia> @time using JuliaDB
[ Info: Precompiling JuliaDB [a93385a2-3734-596a-9a66-3cfbb77141e6]
 40.763768 seconds (11.41 M allocations: 578.145 MiB, 0.84% gc time)

julia> @time using Plots
[ Info: Precompiling Plots [91a5bcdd-55d7-5caf-9e0b-520d859cae80]
 38.993522 seconds (11.61 M allocations: 607.393 MiB, 0.98% gc time)

julia> @time plot(rand(10))
 18.341978 seconds (49.20 M allocations: 2.386 GiB, 8.81% gc time)

Master:

Sysimage built. Summary:
Total ───────  72.745255 seconds
Base: ───────  24.954311 seconds 34.3037%
Stdlibs: ────  47.789192 seconds 65.6939%

julia> @time using DataFrames
[ Info: Recompiling stale cache file /home/keno/.julia/compiled/v1.2/DataFrames/AR9oZ.ji for DataFrames [a93c6f00-e57d-5684-b7b6-d8193f3e46c0]
 24.147537 seconds (10.01 M allocations: 530.951 MiB, 1.24% gc time)

julia> @time using JuliaDB
[ Info: Precompiling JuliaDB [a93385a2-3734-596a-9a66-3cfbb77141e6]
 40.927658 seconds (12.57 M allocations: 631.328 MiB, 0.67% gc time)

julia> @time using Plots
[ Info: Precompiling Plots [91a5bcdd-55d7-5caf-9e0b-520d859cae80]
 39.142164 seconds (12.60 M allocations: 652.750 MiB, 0.91% gc time)

julia> @time plot(rand(10))
 20.319364 seconds (53.89 M allocations: 2.655 GiB, 9.39% gc time)

    JULIA test/subarray
Test  (Worker) | Time (s) | GC (s) | GC % | Alloc (MB) | RSS (MB)
running testset subarray...
subarray   (1) |   194.41 |  16.12 |  8.3 |   25070.70 |   753.57

Test Summary: | Pass  Total
  Overall     |  281    281
    SUCCESS

real    3m22.898s
user    3m21.341s
sys 0m3.514s
```