@Keno 2019-01-23
I'm seeing inference quality regressions with this PR in XLA.jl :(. Looking into it.