@Keno 2019-02-04
Found and fixed the regressions. Good to go from my point of view.