@Keno 2019-02-04
We discussed this PR offline earlier and the conclusion seemed to be that this was the right direction for inference in general, so we'd like to go ahead with it enabled by default.