@vtjnash 2019-02-05
action items we said were:
 - record some basic statistics about how frequently this gets used / how much extra work it incurs (add some global counters to inference, and read out their values later and compute some basic statistics)
 - run nanosoldier
 - add back in profitability limits (like we have for Const, but are missing for PartialTuple)

Seems like maybe we should have more than a single test for it? Upon reflection, PartialTuple probably should have had some more unit tests originally also (some for tmerge, ⊑, some tfuncs, etc.).