@Keno 2019-02-05
If you wanna work on the counters (since you requested that), I can write some tests for the tfuncs that I had to fix up to make this not cause inference regressions.