@Keno 2019-02-13
I've added the profitability limits and some tests. Getting the statistics about what this does to const prop now and then we should be able to merge after a nanosolider run.