@Keno 2019-02-13
The numbers vary a bit run-to run, but I'm getting about an 8% increase in the number of methods for which constant prop happens with this change (things I tried were sysimage and getting to first plot without precompile).