@nanosoldier 2019-02-14
[Your benchmark job](https://github.com/JuliaLang/julia/pull/30783#issuecomment-463411839) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/f77b3842ee811e54ea2d5012175fb3f625e9d23e/a80418c_vs_38d6247/report.md). cc @ararslan