@KristofferC 2019-02-14
No need, it is known to vary in memory usage. We should probably just delete that benchmark.